Fix obtaining server ip address
The previous approach did not work very well if there were multiple IP
interfaces on the system, or if the app was running from a Docker
container. Instead, the server IP address is now obtained via sending a
HTTP GET request to external services which will return the public IP
address back. Also:
- Slight fix in README.md